+++ requirements-dev.txt
@@ -1,3 +1,4 @@
 -r requirements.txt
 mock==2.0.0
 ipython==5.1.0
+pep8==1.7.0
